+++ doc/CHANGES.md
@@ -14,9 +14,10 @@
 
 ## Highlighted Features
 
+* Reduce risk of duplicate request urls (Liz Conlan).
 * Improve speed of the 'old unclassified' requests query by adding a cached
   field to InfoRequest to keep track of when the last public response was
-  made (Liz Conlan)
+  made (Liz Conlan).
 * Various major design and markup improvements to the layout, home page and
   request page (Martin Wright).
 * Adds basic opt-in two factor authentication. Enable it globally with